Create README.md
This is a front-end mini-project built with Next.js and Bootstrap. It is fed from a news API. 
Please note that the news API is free but it can be called more than 10 times. Based on a useEffect callback function the API will be called each time you change either the language or the news category.